tests/wpt/tests/workers/semantics/xhr/001.html

<!doctype html>
<title>async XMLHttpRequest in dedicated worker</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id=log></div>
<script>
var t = async_test();
function runtest() {
  var worker = new Worker('001.js');
  worker.onmessage = this.step_func(function(e) {
    assert_equals(e.data, '');
    this.done();
  });
}
</script>
<iframe src=001-1.xml onload="t.step(runtest);"></iframe>
